Added a ListenerManager
