+++ galasa-simplatform-application/src/main/java/dev/galasa/simplatform/listener/CreditAccountListener.java
@@ -1,5 +1,5 @@
 /*
- * Licensed Materials - Property of IBM
+ * Licensed Materials - Property of IBM 
  * 
  * (c) Copyright IBM Corp. 2019.
  */
@@ -21,23 +21,19 @@ import dev.galasa.simplatform.exceptions.AccountNotFoundException;
 import dev.galasa.simplatform.exceptions.InsufficientBalanceException;
 
 /*
- * 	AccountTransferListener extends from ListenerManager
+ * 	CreditAccountListener extends from ListenerManager
  * 	This 'Listener' is a sub-listener called upon from 
  * 	ListenerManager.
  */
 
-public class AccountTransferListener extends ListenerManager {
+public class CreditAccountListener extends ListenerManager {
 	
 	private ListenerManager manager;
 	
 	private String payload = new String();
-	public String sourceAccountNumber;
-	public String targetAccountNumber;
-
-	public Double value;
-	private String sourceOrgBalance;
-	private String targetOrgBalance;
-
+	private String accountNumber;
+	private double value;
+	
 	private Logger log = Logger.getLogger("Simplatform");
 
     public void sendRequest(ListenerManager manager, String payload) {
@@ -53,31 +49,25 @@ public class AccountTransferListener extends ListenerManager {
         try {
             try {
             	/* Attempts to parse the key information from XML */
-            	parseRequest();
-            	
+                parseRequest();
+                
             } catch (Exception e) {
                 log.warning("Exception found while reading request" + " returning 500");
                 manager.return500();
                 return;
             }
-            
-            /* Sets the balances of the two account numbers prior to processing */
-        	sourceOrgBalance = Double.toString(new Bank().getBalance(sourceAccountNumber));
-        	targetOrgBalance = Double.toString(new Bank().getBalance(targetAccountNumber));
-        	
+
             try {
-            	/* Attempts to transfer the given money from source to target */
-                transferAccounts();
+            	/* Attempts to update the account depending on the request */
+                updateAccount();
                 
             } catch (InsufficientBalanceException e) {
-            	/* Source account doesn't have a large enough balance to transfer */
                 log.warning("Account did not have adequate balance" + " returning 400");
                 manager.return400();
                 return;
                 
             } catch (AccountNotFoundException e) {
-            	/* The account associated with the transfer do not exist... */
-                log.warning("One or more accounts did not exist!" + " returning 400");
+                log.warning("Account did not exist" + " returning 400");
                 manager.return400();
                 return;
                 
@@ -94,41 +84,36 @@ public class AccountTransferListener extends ListenerManager {
     }
     
     /* Creates the XML response payload ready for manager to process */
-	private String formResponse() throws AccountNotFoundException {
-		/* Retrieve new balances of both accounts */
-		String sourceBalance = Double.toString(new Bank().getBalance(sourceAccountNumber));
-		String targetBalance = Double.toString(new Bank().getBalance(targetAccountNumber));
-
+    private String formResponse() throws AccountNotFoundException {
+    	/* Sets the new balance of the account */
+        String balance = Double.toString(new Bank().getBalance(accountNumber));
+        
 		/* Formats XML Response Doc */
-		String xmlText = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-				+ "<SOAP-ENV:Envelope xmlns:soapenv=\"http://schemas.xmlsoap.org/soap/envelope/\" xmlns:SOAP-ENV=\"http://schemas.xmlsoap.org/soap/envelope/\">\n"
-				+ "<SOAP-ENV:Body>\n" + "<TRFACCTOperationResponse xmlns=\"http://www.TRFACCT.Account.Response.com\">\n"
-				+ "<transfer_account_record_response>\n" + "<account_data>\n" + "<account_source_prev_bal>"
-				+ sourceOrgBalance + "</account_source_prev_bal>\n" + "<account_source_new_bal>" + sourceBalance
-				+ "</account_source_new_bal>\n" + "<account_target_prev_bal>" + targetOrgBalance
-				+ "</account_target_prev_bal>\n" + "<account_target_new_bal>" + targetBalance
-				+ "</account_target_new_bal>\n" + "</account_data>\n" + "</transfer_account_record_response>\n"
-				+ "</TRFACCTOperationResponse>\n" + "</SOAP-ENV:Body>\n" + "</SOAP-ENV:Envelope>";
+        String xmlText = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
+                + "<SOAP-ENV:Envelope xmlns:soapenv=\"http://schemas.xmlsoap.org/soap/envelope/\" xmlns:SOAP-ENV=\"http://schemas.xmlsoap.org/soap/envelope/\">\n"
+                + "<SOAP-ENV:Body>\n"
+                + "<UPDACCTOperationResponse xmlns=\"http://www.UPDACCT.Account.Response.com\">\n"
+                + "<update_account_record_response>\n" + "<account_data>\n" + "<account_available_balance>"
+                + balance + "</account_available_balance>\n" + "<account_actual_balance>" + balance
+                + "</account_actual_balance>\n" + "</account_data>\n" + "</update_account_record_response>\n"
+                + "</UPDACCTOperationResponse>\n" + "</SOAP-ENV:Body>\n" + "</SOAP-ENV:Envelope>";
+        
+        return xmlText;
+    }
 
-		return xmlText;
-	}
-
-    private void transferAccounts() throws InsufficientBalanceException, AccountNotFoundException {
-		/* Initialise Bank */
-		Bank bank = new Bank();
-		
-		/* Transfer value from source to target account */
-		bank.transferMoney(sourceAccountNumber, targetAccountNumber, value);
+    /* Credits the given account with the value requested */
+    private void updateAccount() throws InsufficientBalanceException, AccountNotFoundException {
+        new Bank().creditAccount(accountNumber, value);
     }
 
     private void parseRequest() throws Exception {
         DocumentBuilderFactory builderFactory = DocumentBuilderFactory.newInstance();
         DocumentBuilder builder = null;
         Document doc = null;
-        
+
         builder = builderFactory.newDocumentBuilder();
         doc = builder.parse(new ByteArrayInputStream(payload.getBytes()));
-        
+
         /* Finding different elements within the XML Request */
         Element soapEnvelope = doc.getDocumentElement();
         Node soapBody = soapEnvelope.getFirstChild();
@@ -137,9 +122,8 @@ public class AccountTransferListener extends ListenerManager {
         Node accountkey = accountRecord.getFirstChild();
         NodeList dataItems = accountkey.getChildNodes();
         
-        /* Extract and store the account numbers from the XML data set */
-        sourceAccountNumber = dataItems.item(0).getFirstChild().getNodeValue();
-        targetAccountNumber = dataItems.item(1).getFirstChild().getNodeValue();
+        /* Extracting the key information from the XML Document */
+        accountNumber = dataItems.item(1).getFirstChild().getNodeValue();
         Node accountChange = accountkey.getNextSibling();
         String accountValue = accountChange.getFirstChild().getNodeValue();
         value = Double.parseDouble(accountValue);
